Refactoring the shuffler to be a randomizer
