+++ MontyHallKata/Models/Randomizer/IRandomizer.cs
@@ -1,0 +1,7 @@
+namespace MontyHallKata.Models.Randomizer
+{
+    public interface IRandomizer
+    {
+        public T[] GetRandomizedArray<T>(T[] array);
+    }
+}
