--- MontyHallKata/Models/Shuffler/IShuffle.cs
@@ -1,7 +1,0 @@
-namespace MontyHallKata
-{
-    public interface IShuffle
-    {
-        public T[] GetShuffledArray<T>(T[] array);
-    }
-}
